log: Auto wrap baseplate.Error in logs
A pain point currently is that thrift generated baseplate.Error doesn't
provide human readable message automatically since we are using optional
fields, logging it will not produce human readable/meaningful info.

Although we provided WrapBaseplateError helper function, it has its own
limitations, for example it's usually only safe to be used
unconditionally on the client side, which limited its usefulness on the
server side.

Add a wrapper to zapcore to auto wrap baseplate.Errors, so whenever they
are logged via kv pairs they are always auto wrapped.
